Support for start and enddates of tweets

+++ tweet2picasa.py
@@ -1,5 +1,9 @@
 #!/usr/bin/python2.7
 
+''' Takes tweets with Twitpic images and posts them to Picasa '''
+
+__author__ = 'tweet2picasa@kasa.nu'
+
 import twitter
 import re
 import os
@@ -8,10 +12,10 @@ import gdata.photos.service
 import gdata.media
 import gdata.geo
 import argparse
+import time, datetime
+import email.utils
 
-
-
-class PicasaUpload:
+class PicasaUploader:
 	def __init__(self, username, password, title):
 		self.gd_client = gdata.photos.service.PhotosService()
 		self.gd_client.email = username
@@ -26,43 +30,40 @@ class PicasaUpload:
 		photo = self.gd_client.InsertPhotoSimple(self.album_url, 'photo', 
 		    text, filename, content_type='image/jpeg')
 
-
-class TweetFetcher:
-	def get_tweets(self, user, tweets):
-		api = twitter.Api()
-		return api.GetUserTimeline(user, tweets)
-
-
 def main():
-	parser = argparse.ArgumentParser(description='Post all twitpic tweets to Picasa')
+	parser = argparse.ArgumentParser(description='Post all twitpic tweets to Picasa. Only supports public timelines and can only handle max 200 tweets back in time.')
 	parser.add_argument('user_twitter', metavar='username_twitter', help='Username for Twitter')
-	parser.add_argument('tweets', metavar='tweets', type=int, help='Number of tweets to get')
+	parser.add_argument('start_time', metavar='start_time', help='Starttime of tweets (inclusive) in format YYYYmmdd')
+	parser.add_argument('end_time', metavar='end_time', help='Endtime of tweets (exclusive) in format YYYYmmdd')
 	parser.add_argument('user_picasa', metavar='username_picasa', help='Username for Picasa')
 	parser.add_argument('pwd_picasa', metavar='password_picasa', help='Password for Picasa')
 	parser.add_argument('album_name', metavar='album_name', help='Albumname for Picasa')
 
 	args = parser.parse_args()
-	#print args.user_twitter
-
-	statuses = TweetFetcher().get_tweets(args.user_twitter, args.tweets)
 	
-	#pu = PicasaUpload(args.user_picasa, args.pwd_picasa, args.album_name)
+	time_format = "%Y%m%d"
+	startdate = datetime.datetime.fromtimestamp(time.mktime(time.strptime(args.start_time, time_format)))
+	enddate = datetime.datetime.fromtimestamp(time.mktime(time.strptime(args.end_time, time_format)))
+	
+	statuses = twitter.Api().GetUserTimeline(user=args.user_twitter, count=200)
+	
+	pu = PicasaUploader(args.user_picasa, args.pwd_picasa, args.album_name)
 	
 	for s in statuses:
-		m = re.match("(.*)http://twitpic.com/([\d\w]+).*", s.text)
-		if m != None:
-			url = "http://twitpic.com/show/large/" + m.group(2)
-			opener1 = urllib2.build_opener()
-			page1 = opener1.open(url)
-			my_picture = page1.read()
-			filename = ".tweet2picasa_" + m.group(2) + ".jpg"
-			fout = open(filename, "wb")
-			fout.write(my_picture)
-			fout.close()
-			
-			print s.text
-
-			#pu.upload_image(filename, m.group(1))
-			os.remove(filename)
+		tweetdate = datetime.datetime.fromtimestamp(time.mktime(email.utils.parsedate(s.created_at)))
+		if (startdate < tweetdate and tweetdate < enddate):
+			m = re.match("(.*)http://twitddpic.com/([\d\w]+).*", s.text)
+			if m != None:
+				url = "http://twitpic.com/show/large/" + m.group(2)
+				opener1 = urllib2.build_opener()
+				page1 = opener1.open(url)
+				my_picture = page1.read()
+				filename = ".tweet2picasa_" + m.group(2) + ".jpg"
+				fout = open(filename, "wb")
+				fout.write(my_picture)
+				fout.close()
+				
+				pu.upload_image(filename, m.group(1))
+				os.remove(filename)
 
 if __name__ == '__main__': main()
